pyproject.toml
[tool.poetry]
name = "simple-rag-qa"
version = "0.1.0"
description = "Retrieval-Augmented Generation (RAG) pipeline for document retrieval and question answering"
authors = ["Mehrdad Bozorg <bozorgmehrdad@gmail.com>"]
license = "MIT"
packages = [{include = "app"}] 

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.115.12"
uvicorn = "^0.17.0"
faiss-cpu = "^1.7.2"
numpy = "^1.21.0"
pdfplumber = "^0.11.6"
python-multipart = "^0.0.20"
pydantic = "^2.11.1"
sentence-transformers = "^4.0.1"
python-dotenv = "^1.1.0"
openai = "^1.69.0"

[tool.poetry.dev-dependencies]
pytest = "^6.0"